User changes now only bust own Docker build cache.

+++ 01.apt/1.sources/install.sh
@@ -9,10 +9,9 @@ BD=$(dirname $(realpath $0))
 DS=( $BD/default $BD/$(whoami) $BD/$(hostname) )
 TD=/etc/apt/sources.list.d
 
-echo ">>> Apt sources"
 sudo mkdir -pv $TD
 for D in ${DS[@]} ; do
-    if [ -d $D ] ; then
+    if [ -d $D ] && [ ! -z "$(ls -A $D)" ] ; then
         for RF in $(find $D -type f | xargs realpath --relative-to=$D) ; do
             sudo rm -fv $TD/$RF
             # RF must end in a newline for final line to be processed
@@ -22,4 +21,3 @@ for D in ${DS[@]} ; do
         done
     fi
 done
-echo "<<< Apt sources"
